requirements.txt
python-jobspy
pydantic
httpx
PyYAML
python-dotenv
beautifulsoup4
pypdf
pymongo
cryptography
